Rename avrotest to avroregistrytest
