+++ avroregistrytest/avroregistrytest.go
@@ -1,4 +1,4 @@
-package avrotest
+package avroregistrytest
 
 import (
 	"context"
